Add resources for X-Ray 1.5.03

+++ gamedata_cs/shaders/r3/shadow.h
@@ -15,7 +15,7 @@ sampler		smp_jitter;
 //////////////////////////////////////////////////////////////////////////////////////////
 // hardware + PCF
 //////////////////////////////////////////////////////////////////////////////////////////
-half sample_hw_pcf (float4 tc,float4 shift)
+float sample_hw_pcf (float4 tc,float4 shift)
 {
 	static const float 	ts = KERNEL / float(SMAP_size);
 
@@ -27,19 +27,19 @@ half sample_hw_pcf (float4 tc,float4 shift)
 	return s_smap.SampleCmpLevelZero( smp_smap, tc.xy, tc.z).x;
 }
 
-half shadow_hw( float4 tc )
+float shadow_hw( float4 tc )
 {
-  	half	s0		= sample_hw_pcf( tc, float4( -1, -1, 0, 0) );
-  	half	s1		= sample_hw_pcf( tc, float4( +1, -1, 0, 0) );
-  	half	s2		= sample_hw_pcf( tc, float4( -1, +1, 0, 0) );
-  	half	s3		= sample_hw_pcf( tc, float4( +1, +1, 0, 0) );
+  	float	s0		= sample_hw_pcf( tc, float4( -1, -1, 0, 0) );
+  	float	s1		= sample_hw_pcf( tc, float4( +1, -1, 0, 0) );
+  	float	s2		= sample_hw_pcf( tc, float4( -1, +1, 0, 0) );
+  	float	s3		= sample_hw_pcf( tc, float4( +1, +1, 0, 0) );
 
 	return	(s0+s1+s2+s3)/4.h;
 }
 //////////////////////////////////////////////////////////////////////////////////////////
 //	D24X8+PCF
 //////////////////////////////////////////////////////////////////////////////////////////
-half shadow( float4 tc ) 
+float shadow( float4 tc ) 
 {
 	return shadow_hw( tc ); 
 }
@@ -53,11 +53,11 @@ Texture2D	jitter0;
 Texture2D	jitter1;
 //uniform sampler2D	jitter2;
 //uniform sampler2D	jitter3;
-//uniform half4 		jitterS;
+//uniform float4 		jitterS;
 
 Texture2D	jitterMipped;
 
-half4 	test 		(float4 tc, half2 offset)
+float4 	test 		(float4 tc, float2 offset)
 {
 
 //	float4	tcx	= float4 (tc.xy + tc.w*offset, tc.zw);
@@ -68,17 +68,17 @@ half4 	test 		(float4 tc, half2 offset)
 	return s_smap.SampleCmpLevelZero( smp_smap, tc.xy, tc.z).x;
 }
 
-half 	shadowtest 	(float4 tc, float4 tcJ)				// jittered sampling
+float 	shadowtest 	(float4 tc, float4 tcJ)				// jittered sampling
 {
-	half4	r;
+	float4	r;
 
 	const 	float 	scale 	= (2.7f/float(SMAP_size));
 
-//	half4	J0 	= tex2Dproj	(jitter0,tcJ)*scale;
-//	half4	J1 	= tex2Dproj	(jitter1,tcJ)*scale;
+//	float4	J0 	= tex2Dproj	(jitter0,tcJ)*scale;
+//	float4	J1 	= tex2Dproj	(jitter1,tcJ)*scale;
 	tcJ.xy		/=	tcJ.w;
-	half4	J0 	= jitter0.Sample( smp_jitter, tcJ )*scale;
-	half4	J1 	= jitter1.Sample( smp_jitter, tcJ )*scale;
+	float4	J0 	= jitter0.Sample( smp_jitter, tcJ )*scale;
+	float4	J1 	= jitter1.Sample( smp_jitter, tcJ )*scale;
 
 		r.x 	= test 	(tc,J0.xy).x;
 		r.y 	= test 	(tc,J0.wz).y;
@@ -88,16 +88,16 @@ half 	shadowtest 	(float4 tc, float4 tcJ)				// jittered sampling
 	return	dot(r,1.h/4.h);
 }
 
-half 	shadowtest_sun 	(float4 tc, float2 tcJ)			// jittered sampling
+float 	shadowtest_sun 	(float4 tc, float2 tcJ)			// jittered sampling
 {
-	half4	r;
+	float4	r;
 
 //	const 	float 	scale 	= (2.0f/float(SMAP_size));
 	const 	float 	scale 	= (0.7f/float(SMAP_size));
-//	half4	J0 	= tex2D	(jitter0,tcJ)*scale;
-//	half4	J1 	= tex2D	(jitter1,tcJ)*scale;
-	half4	J0 	= jitter0.Sample( smp_jitter, tcJ )*scale;
-	half4	J1 	= jitter1.Sample( smp_jitter, tcJ )*scale;
+//	float4	J0 	= tex2D	(jitter0,tcJ)*scale;
+//	float4	J1 	= tex2D	(jitter1,tcJ)*scale;
+	float4	J0 	= jitter0.Sample( smp_jitter, tcJ )*scale;
+	float4	J1 	= jitter1.Sample( smp_jitter, tcJ )*scale;
 
 		r.x 	= test 	(tc,J0.xy).x;
 		r.y 	= test 	(tc,J0.wz).y;
@@ -107,22 +107,22 @@ half 	shadowtest_sun 	(float4 tc, float2 tcJ)			// jittered sampling
 	return	dot(r,1.h/4.h);
 }
 
-half 	shadow_rain 	(float4 tc, float2 tcJ)			// jittered sampling
+float 	shadow_rain 	(float4 tc, float2 tcJ)			// jittered sampling
 {
-	half4	r;
+	float4	r;
 
 	const 	float 	scale 	= (4.0f/float(SMAP_size));
-//	half4	J0 	= jitter0.Sample( smp_jitter, tcJ )*scale;
-//	half4	J1 	= jitter1.Sample( smp_jitter, tcJ )*scale;
-	half4	J0 	= jitter0.Sample( smp_linear, tcJ )*scale;
-	half4	J1 	= jitter1.Sample( smp_linear, tcJ )*scale;
+//	float4	J0 	= jitter0.Sample( smp_jitter, tcJ )*scale;
+//	float4	J1 	= jitter1.Sample( smp_jitter, tcJ )*scale;
+	float4	J0 	= jitter0.Sample( smp_linear, tcJ )*scale;
+	float4	J1 	= jitter1.Sample( smp_linear, tcJ )*scale;
 
 	r.x 	= test 	(tc,J0.xy).x;
 	r.y 	= test 	(tc,J0.wz).y;
 	r.z		= test	(tc,J1.xy).z;
 	r.w		= test	(tc,J1.wz).x;
 
-//	half4	J0 	= jitterMipped.Sample( smp_base, tcJ )*scale;
+//	float4	J0 	= jitterMipped.Sample( smp_base, tcJ )*scale;
 
 //	r.x 	= test 	(tc,J0.xy).x;
 //	r.y 	= test 	(tc,J0.wz).y;
@@ -135,14 +135,14 @@ half 	shadow_rain 	(float4 tc, float2 tcJ)			// jittered sampling
 //////////////////////////////////////////////////////////////////////////////////////////
 #ifdef  USE_SUNMASK	
 float3x4 m_sunmask;	// ortho-projection
-half sunmask( float4 P )
+float sunmask( float4 P )
 {
 	float2 		tc	= mul( m_sunmask, P );		//
 //	return 		tex2D( s_lmap, tc ).w;			// A8 
 	return 		s_lmap.Sample( smp_linear, tc ).w;	// A8 	
 }
 #else
-half sunmask( float4 P ) { return 1.h; }		// 
+float sunmask( float4 P ) { return 1.h; }		// 
 #endif
 //////////////////////////////////////////////////////////////////////////////////////////
 uniform float4x4	m_shadow;
